+++ apps/gym/src/components/reusable/Logo/DesktopLogo.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link'
 
 export const DesktopLogo = () => {
   return (
-    <Link key={1} className={styles.container} href={DESKTOP_NAV_OPTIONS[0].path}>
+    <Link key={1} className={styles.desktop_container} href={DESKTOP_NAV_OPTIONS[0].path}>
       <img src='/images/dumbbell-blue.png' alt='logo' width={25} height={25} />
       <span className={styles.desktop_brand}>JACKED</span>
     </Link>